fix: remove redundant check for non-integer input
I had a try/catch block for parsing integer input in displayResults, but
in the previous conditional I was also checking for any input that was
not an integer with a regex. Previously this would show all words if a
non-integer was entered and try/catch block was never used. Remove
regex, try/catch block used, and only top 20 words displayed if
non-integer entered.